finish the development of getting raw data.
